requirements.txt
appdirs==1.4.3
asgiref==3.8.1
attrs==18.2.0
babel==2.16.0
black==18.9b0
callouts==0.1.1
certifi==2018.11.29
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Django==5.1.4
django-livereload-server==0.5.1
django-livesync==0.5
django-user-agents==0.4.0
fake-useragent==1.4.0
future==0.17.1
ghp-import==2.1.0
humanfriendly==10.0
idna==2.8
ijson==3.3.0
isort==4.3.4
Jinja2==3.1.5
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-click==0.8.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.49
mkdocs-material-extensions==1.3.1
numpy==2.2.1
oauthlib==3.0.1
packaging==24.2
paginate==0.5.7
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pydantic==1.10.19
Pygments==2.18.0
pymdown-extensions==10.13
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-twitter==3.5
pytz==2023.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
ratelimit==2.2.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.2.0
rich==13.3.5
rich-click==1.6.1
six==1.16.0
soupsieve==1.8
sqlparse==0.5.3
tenacity==8.2.3
toml==0.10.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.2
user-agents==2.2.0
user_agent==0.1.10
watchdog==6.0.0
websocket-client==1.8.0
